# contacts-api/docs/swagger.yaml
basePath: /
definitions:
  contacts.APIError:
    properties:
      error:
        type: string
    type: object
  contacts.Contact:
    properties:
      cpfCnpj:
        type: string
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      updatedAt:
        type: string
    type: object
  contacts.CreateContactRequest:
    properties:
      cpfCnpj:
        type: string
      email:
        type: string
      name:
        minLength: 3
        type: string
      phone:
        type: string
    required:
    - cpfCnpj
    - email
    - name
    - phone
    type: object
  contacts.ListContactsResponse:
    properties:
      contacts:
        items:
          $ref: '#/definitions/contacts.Contact'
        type: array
      page:
        type: integer
      perPage:
        type: integer
      total:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Contacts API built with Go
  title: Contacts API
  version: "1.0"
paths:
  /contacts:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 10)
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contacts.ListContactsResponse'
      summary: List contacts
      tags:
      - Contacts
    post:
      consumes:
      - application/json
      description: Creates a contact with name, email, phone, and CPF/CNPJ
      parameters:
      - description: Contact data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contacts.CreateContactRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/contacts.CreateContactRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/contacts.APIError'
      summary: Create a new contact
      tags:
      - Contacts
  /contacts/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a contact by ID
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/contacts.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/contacts.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/contacts.APIError'
      summary: Delete a contact
      tags:
      - Contacts
swagger: "2.0"